+++ cli/aux/hook.c
@@ -5,6 +5,7 @@
 
 #include "hook.h"
 #include "config.h"
+#include "../tec.h"
 #include "../../lib/libtec.h"
 
 static char pathname[PATH_MAX + 1];
@@ -25,9 +26,13 @@ int hook_action(tec_arg_t * args, char *cmd)
     if (teccfg.opts.hook == false)
         return 0;
 
-    for (; hooks; hooks = hooks->next)
-        if (strcmp(cmd, hooks->cmd) == 0)
-            system(path_pgn(args, hooks->pgname, hooks->pgncmd));
+    for (; hooks; hooks = hooks->next) {
+        if (strcmp(cmd, hooks->cmd) == 0) {
+            char *cmd = path_pgn(args, hooks->pgname, hooks->pgncmd);
+            dlog(1, cmd);
+            system(cmd);
+        }
+    }
     return 0;
 }
 
@@ -94,3 +99,33 @@ char *hook_list(struct tec_hook *hooks, char *pgnout, char *project, char *task)
     return pgnout;
 }
 */
+
+/* TODO:
+    Structure: tec PGN -i -b -p COMMAND [OPTION]... [APRS]
+    1. Can't use getopt cuz there might be different options for plugins and
+       their commads.
+    2. Use for/while loop to circle through options and their arguments.
+    3. Separate plugin options from plugin command options.
+    4. Or maybe it's better to let the plugin to handle plugin options and the rest.
+    5. Add a better parser to include all the other tec options like -C, -H, -F, etc.
+*/
+int tec_cli_plugin(int argc, char **argv, tec_ctx_t * ctx)
+{
+    int i;
+    char *pgn;
+    char pgnexec[BUFSIZ + 1] = { 0 };
+
+    i = 0;
+    pgn = argv[i++];
+
+    sprintf(pgnexec, "%s/%s/%s -T %s -P %s ", teccfg.base.pgn, pgn, pgn,
+            teccfg.base.task, teccfg.base.pgn);
+
+    for (; i < argc; ++i) {
+        strcat(pgnexec, argv[i]);
+        strcat(pgnexec, " ");
+    }
+
+    dlog(1, "pgnexec: %s", pgnexec);
+    return system(pgnexec);
+}